+++ helpers/include/edhoc_cipher_suite_2.h
@@ -1,5 +1,5 @@
 /**
- * \file    cipher_suite_2.h
+ * \file    edhoc_cipher_suite_2.h
  * \author  Kamil Kielbasa
  * \brief   Cipher suite 2 contains:
  *            - AEAD algorithm                      = AES-CCM-16-64-128
@@ -16,8 +16,8 @@
  */
 
 /* Header guard ------------------------------------------------------------ */
-#ifndef CIPHER_SUITE_2_H
-#define CIPHER_SUITE_2_H
+#ifndef EDHOC_CIPHER_SUITE_2_H
+#define EDHOC_CIPHER_SUITE_2_H
 
 /* Include files ----------------------------------------------------------- */
 
@@ -40,6 +40,17 @@
 
 /* Module types and type definitiones -------------------------------------- */
 /* Module interface variables and constants -------------------------------- */
+
+/**
+ * \brief Get EDHOC crypto structure for cipher suite 2.
+ */
+const struct edhoc_crypto *edhoc_cipher_suite_2_get_crypto(void);
+
+/**
+ * \brief Get EDHOC keys structure for cipher suite 2.
+ */
+const struct edhoc_keys *edhoc_cipher_suite_2_get_keys(void);
+
 /* Static variables and constants ------------------------------------------ */
 /* Static function declarations -------------------------------------------- */
 /* Static function definitions --------------------------------------------- */
@@ -48,19 +59,19 @@
 /**
  * \brief Crypto key generation.
  */
-int cipher_suite_2_key_import(void *user_context, enum edhoc_key_type key_type,
+int edhoc_cipher_suite_2_key_import(void *user_context, enum edhoc_key_type key_type,
 			      const uint8_t *raw_key, size_t raw_key_len,
 			      void *kid);
 
 /**
  * \brief Crypto key destroy.
  */
-int cipher_suite_2_key_destroy(void *user_context, void *kid);
+int edhoc_cipher_suite_2_key_destroy(void *user_context, void *kid);
 
 /**
  * \brief ECDH make key pair.
  */
-int cipher_suite_2_make_key_pair(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_make_key_pair(void *user_context, const void *key_id,
 				 uint8_t *restrict private_key,
 				 size_t private_key_size,
 				 size_t *restrict private_key_length,
@@ -71,7 +82,7 @@ int cipher_suite_2_make_key_pair(void *user_context, const void *key_id,
 /**
  * \brief ECDH key agreement.
  */
-int cipher_suite_2_key_agreement(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_key_agreement(void *user_context, const void *key_id,
 				 const uint8_t *peer_public_key,
 				 size_t peer_public_key_length,
 				 uint8_t *shared_secret,
@@ -81,7 +92,7 @@ int cipher_suite_2_key_agreement(void *user_context, const void *key_id,
 /**
  * \brief ECDSA signature.
  */
-int cipher_suite_2_signature(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_signature(void *user_context, const void *key_id,
 			     const uint8_t *input, size_t input_length,
 			     uint8_t *signature, size_t signature_size,
 			     size_t *signature_length);
@@ -89,14 +100,14 @@ int cipher_suite_2_signature(void *user_context, const void *key_id,
 /**
  * \brief ECDSA signature verification.
  */
-int cipher_suite_2_verify(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_verify(void *user_context, const void *key_id,
 			  const uint8_t *input, size_t input_length,
 			  const uint8_t *signature, size_t signature_length);
 
 /**
  * \brief HKDF extract.
  */
-int cipher_suite_2_extract(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_extract(void *user_context, const void *key_id,
 			   const uint8_t *salt, size_t salt_len,
 			   uint8_t *psuedo_random_key,
 			   size_t psuedo_random_key_size,
@@ -105,7 +116,7 @@ int cipher_suite_2_extract(void *user_context, const void *key_id,
 /**
  * \brief HKDF expand.
  */
-int cipher_suite_2_expand(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_expand(void *user_context, const void *key_id,
 			  const uint8_t *info, size_t info_length,
 			  uint8_t *output_keying_material,
 			  size_t output_keying_material_length);
@@ -113,7 +124,7 @@ int cipher_suite_2_expand(void *user_context, const void *key_id,
 /**
  * \brief AEAD encrypt.
  */
-int cipher_suite_2_encrypt(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_encrypt(void *user_context, const void *key_id,
 			   const uint8_t *nonce, size_t nonce_length,
 			   const uint8_t *additional_data,
 			   size_t additional_data_length,
@@ -124,7 +135,7 @@ int cipher_suite_2_encrypt(void *user_context, const void *key_id,
 /**
  * \brief AEAD decrypt.
  */
-int cipher_suite_2_decrypt(void *user_context, const void *key_id,
+int edhoc_cipher_suite_2_decrypt(void *user_context, const void *key_id,
 			   const uint8_t *nonce, size_t nonce_length,
 			   const uint8_t *additional_data,
 			   size_t additional_data_length,
@@ -135,8 +146,9 @@ int cipher_suite_2_decrypt(void *user_context, const void *key_id,
 /**
  * \brief Hash function.
  */
-int cipher_suite_2_hash(void *user_context, const uint8_t *input,
+int edhoc_cipher_suite_2_hash(void *user_context, const uint8_t *input,
 			size_t input_length, uint8_t *hash, size_t hash_size,
 			size_t *hash_length);
 
-#endif /* CIPHER_SUITE_2_H */
+#endif /* EDHOC_CIPHER_SUITE_2_H */
+