+++ library/edhoc_message_3.c
@@ -11,10 +11,16 @@
 
 /* Include files ----------------------------------------------------------- */
 
+#ifdef __ZEPHYR__
+#include <zephyr/logging/log.h>
+LOG_MODULE_DECLARE(libedhoc, CONFIG_LIBEDHOC_LOG_LEVEL);
+#endif
+
 /* EDHOC header: */
 #define EDHOC_ALLOW_PRIVATE_ACCESS
 #include "edhoc.h"
 #include "edhoc_common.h"
+#include "edhoc_log.h"
 
 /* Standard library headers: */
 #include <stdint.h>